requirements.txt
anyascii==0.3.2
asgiref==3.7.2
# backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.10
django-appconf==1.0.6
django-classy-tags==4.1.0
django-cms==4.1.0
django-entangled==0.5.4
django-extensions==3.2.3
django-filer==3.1.1
django-filter==23.5
django-formtools==2.5.1
django-fsm==2.8.1
django-jazzmin==2.6.0
django-modelcluster==6.2.1
django-modeltranslation==0.18.11
django-parler==2.3
django-permissionedforms==0.1
django-polymorphic==3.1.0
django-rosetta==0.10.0
django-sekizai==4.1.0
django-select2==8.1.2
django-taggit==4.0.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.0
djangocms-alias==2.0.0
djangocms-attributes-field==3.0.0
djangocms-file==3.0.1
djangocms-frontend==1.2.2
djangocms-googlemap==2.2.0
djangocms-picture==4.1.1
djangocms-snippet==3.1.1
djangocms-style==3.1.0
djangocms-text-ckeditor==5.1.5
djangocms-versioning==2.0.0
djangocms-video==3.1.0
djangorestframework==3.14.0
draftjs-exporter==2.1.7
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.6
jsonfield==3.1.0
l18n==2021.3
lxml==5.1.0
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
pillow-heif==0.14.0
polib==1.2.0
#psycopg2==2.9.9
python-gettext==5.0
pytz==2023.3.post1
reportlab==4.1.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
telepath==0.3.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.1.0
user-agents==2.2.0
# wagtail==5.2.2
webencodings==0.5.1
whitenoise==6.6.0
Willow==1.6.3
